requirements.txt
absl-py==1.4.0
anyascii==0.3.2
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
attrs==22.2.0
av==10.0.0
backcall==0.2.0
beautifulsoup4==4.11.2
bleach==6.0.0
boto3==1.26.86
botocore==1.29.86
cachetools==5.3.0
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comm==0.1.2
contourpy==1.0.7
contractions==0.1.73
cycler==0.11.0
Cython==0.29.33
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
easydict==1.10
einops==0.6.0
executing==1.2.0
fastjsonschema==2.16.3
filelock==3.10.6
fonttools==4.39.3
fqdn==1.5.1
future==0.18.3
fvcore==0.1.5.post20221221
gensim==3.8.1
google-auth==2.17.3
google-auth-oauthlib==1.0.0
grpcio==1.54.0
h5py==3.8.0
HyperPyYAML==1.2.0
idna==3.4
importlib-metadata==6.1.0
importlib-resources==5.12.0
iopath==0.1.10
ipykernel==6.21.3
ipython==8.11.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter_client==8.0.3
jupyter_core==5.2.0
jupyter_server==2.4.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
kiwisolver==1.4.4
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==2.0.5
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186064646/work
mkl-service==2.4.0
nbclassic==0.5.3
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
notebook==6.5.3
notebook_shim==0.2.2
numpy==1.24.2
oauthlib==3.2.2
opencv-python==4.7.0.72
packaging==23.0
pandas==2.0.0
pandocfilters==1.5.0
parameterized==0.9.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pkgutil_resolve_name==1.3.10
platformdirs==3.1.0
portalocker==2.7.0
prettytable==3.6.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==4.22.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyahocorasick==2.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.14.0
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-json-logger==2.0.7
pytorchvideo==0.1.5
pytz==2023.3
PyYAML==6.0
pyzmq==25.0.0
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.6.0
sacremoses==0.0.53
scikit-learn==1.2.2
scipy==1.10.1
Send2Trash==1.8.0
sentencepiece==0.1.97
six @ file:///tmp/build/80754af9/six_1644875935023/work
smart-open==6.3.0
sniffio==1.3.0
soupsieve==2.4
stack-data==0.6.2
tabulate==0.9.0
tensorboard==2.12.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
termcolor==2.2.0
terminado==0.17.1
textsearch==0.0.24
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tinycss2==1.2.1
tokenizers==0.9.4
torch==1.6.0
torchaudio==0.6.0
torchvision==0.7.0
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.0.0
typing_extensions==4.5.0
tzdata==2023.3
uri-template==1.2.0
urllib3==1.26.15
validators==0.20.0
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.2.3
wget==3.2
widgetsnbextension==3.5.2
yacs==0.1.8
zipp==3.15.0
